+++ src/DeploymentsManager.ts
@@ -96,7 +96,7 @@ export class DeploymentsManager {
       fixtureCounter: 0,
       snapshotCounter: 0,
       pastFixtures: {},
-      logEnabled: false,
+      logEnabled: process.env['HARDHAT_DEPLOY_LOG'] ? true : false,
       pendingTransactions: {},
       savePendingTx: false,
       gasPrice: undefined,
@@ -619,6 +619,11 @@ export class DeploymentsManager {
     }
     if (typeof deployment.abi === undefined) {
       throw new Error('deployment need an ABI');
+    }
+    if (name.includes('/') || name.includes(':')) {
+      throw new Error(
+        `deployment name must not be a path or Fully Qualified Name - for such purposes consider using the "contract" field of deployment options`
+      );
     }
 
     const chainId = await this.getChainId();