+++ src/etherscan.ts
@@ -24,6 +24,10 @@ function logSuccess(...args: any[]) {
   console.log(chalk.green(...args));
 }
 
+function sleep(ms: number) {
+  return new Promise((resolve) => setTimeout(resolve, ms));
+}
+
 function extractOneLicenseFromSourceFile(source: string): string | undefined {
   const licenses = extractLicenseFromSources(source);
   if (licenses.length === 0) {
@@ -101,6 +105,7 @@ export async function submitSources(
     license?: string;
     fallbackOnSolcInput?: boolean;
     forceLicense?: boolean;
+    sleepBetween?: boolean;
   }
 ): Promise<void> {
   config = config || {};
@@ -108,6 +113,7 @@ export async function submitSources(
   const licenseOption = config.license;
   const forceLicense = config.forceLicense;
   const etherscanApiKey = config.etherscanApiKey;
+  const sleepBetween = config.sleepBetween;
   const chainId = await hre.getChainId();
   const all = await hre.deployments.all();
   let host: string;
@@ -135,6 +141,9 @@ export async function submitSources(
       break;
     case '128':
       host = 'https://api.hecoinfo.com';
+      break;
+    case '137':
+      host = "https://api.polygonscan.com";
       break;
     case '256':
       host = 'https://api-testnet.hecoinfo.com';
@@ -320,7 +329,7 @@ export async function submitSources(
       guid = submissionData.result;
     } else {
       logError(
-        `contract ${name} failed to submit : "${submissionData.message}"`,
+        `contract ${name} failed to submit : "${submissionData.message}" : "${submissionData.result}"`,
         submissionData
       );
       return;
@@ -403,5 +412,10 @@ export async function submitSources(
 
   for (const name of Object.keys(all)) {
     await submit(name);
+
+    if (sleepBetween) {
+      // sleep between each verification so we don't exceed the API rate limit
+      await sleep(500);
+    }
   }
 }