+++ src/helpers.ts
@@ -710,6 +710,7 @@ export function addHelpers(
             };
           } else {
             result = deployment as DeployResult;
+            result.newlyDeployed = false;
           }
         } else {
           if (!diffResult.address) {
@@ -737,7 +738,9 @@ export function addHelpers(
             newlyDeployed: false,
           };
         }
-        log(`reusing "${name}" at ${result.address}`);
+        if (options.log) {
+          log(`reusing "${name}" at ${result.address}`);
+        }
       }
     } else {
       result = await _deploy(name, options);
@@ -1221,11 +1224,13 @@ Plus they are only used when the contract is meant to be used as standalone when
               LedgerSigner = hardwareWalletModule.LedgerSigner;
             }
             ethersSigner = new LedgerSigner(provider);
+            hardwareWallet = 'ledger';
           } else if (registeredProtocol.startsWith('privatekey')) {
+            ethersSigner = new Wallet(registeredProtocol.substr(13), provider);
+          } else if (registeredProtocol.startsWith('gnosis')) {
             ethersSigner = new Wallet(registeredProtocol.substr(13), provider);
           }
         }
-        hardwareWallet = 'ledger';
       }
     }